(fead): add new hooks
