+++ design_system/components/molecules/HelloWave/HelloWave.tsx
@@ -1,19 +1,10 @@
-import { useEffect } from 'react';
-import Animated, { useSharedValue, useAnimatedStyle } from 'react-native-reanimated';
+import Animated from 'react-native-reanimated';
 import { ThemedText } from '@/design_system/components/atoms';
-import { waveAnimation } from '@/design_system/ui/animations';
 import { styles } from './HelloWave.styles';
+import { useHelloWaveAnimation } from '@/hooks/components/useHelloWaveAnimation';
 
 export function HelloWave() {
-  const rotationAnimation = useSharedValue(0);
-
-  useEffect(() => {
-    waveAnimation.rotate(rotationAnimation);
-  }, []);
-
-  const animatedStyle = useAnimatedStyle(() => 
-    waveAnimation.style(rotationAnimation)
-  );
+  const animatedStyle = useHelloWaveAnimation(); 
 
   return (
     <Animated.View style={animatedStyle}>
