+++ design_system/components/molecules/MessageBubble/MessageBubble.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { View, Alert, TouchableWithoutFeedback } from 'react-native';
+import { View, TouchableWithoutFeedback } from 'react-native';
 import { ThemedText } from '@/design_system/components/atoms';
+import { styles } from './MessageBubble.styles';
+import { useMessageBubble } from '@/hooks/components/useMessageBubble'; // Import the custom hook
 import { Message } from '@/hooks/useChats';
-import { useColorScheme } from '@/hooks/useColorScheme';
-import { styles, getBubbleColors } from './MessageBubble.styles';
 
 interface MessageBubbleProps {
   message: Message;
@@ -12,22 +12,7 @@ interface MessageBubbleProps {
 }
 
 export function MessageBubble({ message, isCurrentUser, onDeleteMessage }: MessageBubbleProps) {
-  const colorScheme = useColorScheme();
-  const isDark = colorScheme === 'dark';
-  const bubbleColors = getBubbleColors(isDark, isCurrentUser);
-
-  const handleLongPress = () => {
-    if (isCurrentUser && onDeleteMessage) {
-      Alert.alert(
-        'Delete Message',
-        'Do you want to delete this message?',
-        [
-          { text: 'Cancel', style: 'cancel' },
-          { text: 'Delete', onPress: () => onDeleteMessage(message.id), style: 'destructive' }
-        ]
-      );
-    }
-  };
+  const { isDark, bubbleColors, handleLongPress } = useMessageBubble({ message, isCurrentUser, onDeleteMessage });
 
   const formatTime = (timestamp: number) => {
     return new Date(timestamp).toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
